create PageList component
