+++ src/components/collection-log/index.ts
@@ -1,0 +1,3 @@
+import PageList from './PageList';
+
+export { PageList };
